Avoid runs existence check for new jobs
The runs are now created without checking for their existence in the  DB.
This is safe as the DB model prevents these runs from existing before
the corresponding jobs exists. This is indeed a great optimization when
importing a lot of new jobs, typically on new clusters initial import.

The runs have been made an attribute of Job class in order to avoid a
complex (and compute intensive) search in JobImporter update() method.

Fix #8