Report new error E_E0001 instead of Exception
When EventImporterSlurm founds an event associated to a node unknown in
cluster architecture, HPCStats now reports a new error E_E0001 (that
can be ignored, as well as all errors managed by the error manager)
instead of raising a source exception.

This commit also updates the tests and the documentation accordingly.