Add optional slurm prefix conf param
This commit introduces the new optional Slurm prefix configuration
parameter. It is used to control the prefix of the table names in
Slurm accounting relational database. By default, this is the cluster
name.

As explained in the documentation, this parameter might only be useful
when someone does not want to use the Slurm cluster name in HPCStats.