Compose fake department name base on primary group
The UserLdapImporter tries to compose the user department name based on
its groups membership, trying to find a group name with a specific
format.

This commit adds a fallback method to this logic: if no group name with
the expected format is found, the importer consider the primary group
name of the user to be its direction, ie. the first part of its
department name. This primary group name is concatenated to a
configurable default subdirection name.