Add optional primary group aliasing feature
This commit introduces a feature to provided optional aliasing of the
primary group names. In an additional files, it is possible to
associate primary group names with an actual direction name.

The commit also calls the check() method on load() in order to check
the group alias file actually exists before trying to open it. This
check() call removes the need to connect the LDAP directory afterwards.