tests: use native mock_open()
The test suite relied on internally defined mock_open() because we were
initially using python mock library < 0.8.0 but this is prehistory. It
has been a long time since mock library provides its native
implementation of mock_open().

This commit updates the tests to use this native function and removed
internal definition. Unfortunately, the MagicMock returned by
mock_open() does not defines all the methods of Python builtin file
handler, esp. iterations, which are notably used by csv library. That
is why also explicitely define the return value of the __iter__ method.

There is a very interesting read at this URL on this subject:
https://stackoverflow.com/questions/24779893/customizing-unittest-mock-mock-open-for-iteration/24779923#24779923