Just log warning when job node is unknown
HPCStats JobImporterSlurm used to raise a fatal HPCStatsSourceError
exception when a job was allocated a node not found in loaded
architecture. This could notably happen when not are renamed on
cluster. With this commit, the exception is replaced by a simple
warning log message. The corresponding run is simply not created in
HPCStats database. The log message can optionally be downgraded to
debug level if the associated error code E_J0006 is set ignored in
configuration file.

This commit also updates the corresponding unit test accordingly.