Correct handling of unlimited walltime
A walltime set as unlimited ("-1") is stored as 2^32-1 in unsigned
cluster_job_table.timelimit field. The job.job_walltime interval goes
up to 178M years but accepts only 2^31-1 seconds (68 years) on input
when the (implied) unit is second. This modification considers any
value above it to mean "unlimited" and stores it as such ("-1").